requirements.txt
# ── Core ML ──
torch
torchvision
Pillow
numpy
scikit-learn

# ── Visualization & Logging ──
matplotlib
seaborn
tqdm
tensorboard

# ── API (for later phases) ──
fastapi
uvicorn
python-multipart

# ── Optional: ONNX export ──
onnx
onnxruntime
